requirements.txt
Django>=4.2
 djangorestframework>=3.14
 djangorestframework-authtoken>=1.1
 PyPDF2>=3.0
 sentence-transformers>=2.2
 transformers>=4.36
 faiss-cpu>=1.7
 numpy>=1.23 